Moved sort functions into generalUtils
